Add separate shellyUpdateHandler
